refactor: 폴더명 및 컴포넌트 파일이름 양식에 맞게 수정 (#88)
