+++ src/components/PaginatedTable/PaginatedTable.tsx
@@ -1,8 +1,8 @@
 import { useState, useEffect } from 'react';
-import Table from '../table/table';
+import Table from '../table/Table';
 import Pagination from '../pagination/pagination';
 import { Modal } from '@/components/modal/Modal';
-import SalarySelect from '@/components/salaryselect/salarySelect';
+import SalarySelect from '@/components/salaryselect/SalarySelect';
 import ModalPortal from '@/components/modal/ModalPortal';
 import { createClient } from '@supabase/supabase-js';
 import EditModal from './EditModal/editModal';
