+++ src/components/schedule-management/schedule-list/admin-schedule-list/AdminScheduleList.styles.ts
