+++ src/pages/register/Register.styles.ts
